make re-entrant and add _bivio_home_env_update for WNAT_WANT_PERL

+++ installers/perl-dev/radiasoft-download.sh
@@ -9,12 +9,18 @@ perl_dev_main() {
     if (( $EUID == 0 )); then
         install_err 'run as vagrant (or other ordinary user), not root'
     fi
-    install -m 600 /dev/stdin ~/.pre_bivio_bashrc <<'EOF'
+    if ! grep -s -q BIVIO_WANT_PERL=1 ~/.pre_bivio_bashrc; then
+        install -m 600 /dev/stdin ~/.pre_bivio_bashrc <<'EOF'
 export BIVIO_WANT_PERL=1
 export BIVIO_HTTPD_PORT=8000
 EOF
-    install_repo_as_root biviosoftware/container-perl
+    fi
+    if ! perl -MGMP::Mpf -e 1 2>&1; then
+        install_repo_as_root biviosoftware/container-perl
+    fi
     set +euo pipefail
+    . ~/.bashrc
+    _bivio_home_env_update -f
     . ~/.bashrc
     bivio sql init_dbms >& /dev/null || true
     # always recreate db
